Add Ollama backend and a new debug notebook for embedding comparison
- Introduced `OllamaSentenceTransformer` as a custom wrapper for the Ollama backend.
- Refactored `sentence_transformer_backend_comparison.py` to support the new backend.
- Created `debug/ollama_comparison.ipynb` for embedding computations and comparisons.
- Moved the Ollama embedding function to `docuverse/utils/embeddings/ollama_embedding_function.py` for cleaner modularity.
